Refactor Plugins check_threshold and check_range
They have now been moved into their own module, called
classic_threshold_syntax

This is in style with 'new_threshold_syntax' being in a
seperate module. and this gives us guideline on where to
put 'relational expressions' a.k.a. NSClient style thresholds.

As before, backwards compatibility is maintained via explicit assignments
in Plugins/__init__.py but later we might add a deprecationwarning here.
