+++ pynag/Plugins/__init__.py
@@ -29,13 +29,13 @@ import sys
 import os
 import traceback
 import signal
-from platform import node
 from optparse import OptionParser, OptionGroup
-from pynag.Utils import PerfData, reconsile_threshold, runCommand
+from pynag.Utils import PerfData, reconsile_threshold
 from pynag.Parsers import ExtraOptsParser
 import pynag.Utils
 import pynag.errors
 from . import new_threshold_syntax
+from . import classic_threshold_syntax
 
 # Map the return codes
 OK = 0
@@ -268,7 +268,7 @@ class simple(object):
         self._check_messages_and_exit()
 
     def _range_checker(self, value, range_threshold):
-        """ deprecated. Use pynag.Plugins.check_range() """
+        """ deprecated. Use pynag.Plugins.class_threshold_syntax.check_range() """
         return check_range(value=value, range_threshold=range_threshold)
 
     def send_nsca(self, *args, **kwargs):
@@ -396,139 +396,6 @@ class simple(object):
             return self.data[key]
         else:
             return None
-
-
-def check_threshold(value, warning=None, critical=None):
-    """ Checks value against warning/critical and returns Nagios exit code.
-
-    Format of range_threshold is according to:
-    http://nagiosplug.sourceforge.net/developer-guidelines.html#THRESHOLDFORMAT
-
-    Returns (in order of appearance):
-        UNKNOWN int(3)  -- On errors or bad input
-        CRITICAL int(2) -- if value is within critical threshold
-        WARNING int(1)  -- If value is within warning threshold
-        OK int(0)       -- If value is outside both tresholds
-    Arguments:
-        value -- value to check
-        warning -- warning range
-        critical -- critical range
-
-    # Example Usage:
-    >>> check_threshold(88, warning="0:90", critical="0:95")
-    0
-    >>> check_threshold(92, warning=":90", critical=":95")
-    1
-    >>> check_threshold(96, warning=":90", critical=":95")
-    2
-    """
-    if critical and not check_range(value, critical):
-        return CRITICAL
-    elif warning and not check_range(value, warning):
-        return WARNING
-    else:
-        return OK
-
-
-def check_range(value, range_threshold=None):
-    """ Returns True if value is within range_threshold.
-
-    Format of range_threshold is according to:
-    http://nagiosplug.sourceforge.net/developer-guidelines.html#THRESHOLDFORMAT
-
-    Arguments:
-        value -- Numerical value to check (i.e. 70 )
-        range -- Range to compare against (i.e. 0:90 )
-    Returns:
-        True  -- If value is inside the range
-        False -- If value is outside the range (alert if this happens)
-        False -- if invalid value is specified
-
-    Summary from plugin developer guidelines:
-    ---------------------------------------------------------
-    x       Generate an alert if x...
-    ---------------------------------------------------------
-    10  	< 0 or > 10, (outside the range of {0 .. 10})
-    10:     < 10, (outside {10 .. ∞})
-    ~:10    > 10, (outside the range of {-∞ .. 10})
-    10:20   < 10 or > 20, (outside the range of {10 .. 20})
-    @10:20  ≥ 10 and ≤ 20, (inside the range of {10 .. 20})
-    10      < 0 or > 10, (outside the range of {0 .. 10})
-    ---------------------------------------------------------
-
-
-    # Example runs for doctest, False should mean alert
-    >>> check_range(78, "90") # Example disk is 78% full, threshold is 90
-    True
-    >>> check_range(5, 10) # Everything between 0 and 10 is True
-    True
-    >>> check_range(0, 10) # Everything between 0 and 10 is True
-    True
-    >>> check_range(10, 10) # Everything between 0 and 10 is True
-    True
-    >>> check_range(11, 10) # Everything between 0 and 10 is True
-    False
-    >>> check_range(-1, 10) # Everything between 0 and 10 is True
-    False
-    >>> check_range(-1, "~:10") # Everything Below 10
-    True
-    >>> check_range(11, "10:") # Everything above 10 is True
-    True
-    >>> check_range(1, "10:") # Everything above 10 is True
-    False
-    >>> check_range(0, "5:10") # Everything between 5 and 10 is True
-    False
-    >>> check_range(0, "@5:10") # Everything outside 5:10 is True
-    True
-    >>> check_range(None) # Return False if value is not a number
-    False
-    >>> check_range("10000000 PX") # What happens on invalid input
-    False
-    >>> check_range("10000000", "invalid:invalid") # What happens on invalid range
-    Traceback (most recent call last):
-    ...
-    PluginError: Invalid threshold format: invalid:invalid
-    """
-
-    # Return false if value is not a number
-    try:
-        float(value)
-    except Exception:
-        return False
-
-    # if no range_threshold is provided, assume everything is ok
-    if not range_threshold:
-        range_threshold = '~:'
-    range_threshold = str(range_threshold)
-    # If range starts with @, then we do the opposite
-    if range_threshold[0] == '@':
-        return not check_range(value, range_threshold[1:])
-
-    if range_threshold.find(':') > -1:
-        (start, end) = (range_threshold.split(':', 1))
-    # we get here if ":" was not provided in range_threshold
-    else:
-        start = ''
-        end = range_threshold
-    # assume infinity if start is not provided
-    if start == '~':
-        start = None
-    # assume start=0 if start is not provided
-    if start == '':
-        start = 0
-    # assume infinity if end is not provided
-    if end == '':
-        end = None
-
-    try:
-        # start is defined and value is lower than start
-        if start is not None and float(value) < float(start):
-            return False
-        if end is not None and float(value) > float(end):
-            return False
-    except ValueError:
-        raise PluginError("Invalid threshold format: %s" % range_threshold)
-    return True
 
 
 class PluginHelper(object):
@@ -1206,3 +1073,7 @@ class PluginHelper(object):
 
     def __repr__(self):
         return self.get_plugin_output(long_output='', perfdata='')
+
+# For backwards compatibility:
+check_range = classic_threshold_syntax.check_range
+check_threshold = classic_threshold_syntax.check_threshold
